move settings file
